# AI Customer Service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat/send": {
      "post": {
        "tags": [
          "聊天会话管理"
        ],
        "summary": "Send Chat Message",
        "description": "发送一条聊天消息：\n- 如果未传 session_id：自动创建新会话\n- 如果传入 session_id：在现有会话中继续对话\n- 返回：会话ID + 助手的回复消息\n\n功能说明：\n- 智能意图识别（订单查询、创建、修改、删除、退货、投诉、FAQ、闲聊）\n- 情绪识别与安抚性回复\n- 多轮对话上下文记忆\n- RAG 知识库检索（针对商品咨询等专业问题）",
        "operationId": "send_chat_message_api_chat_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChatResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/chat/history": {
      "post": {
        "tags": [
          "聊天会话管理"
        ],
        "summary": "Get Chat History",
        "description": "获取会话历史消息（倒序排列，最多返回 limit 条）\n\n参数说明：\n- session_id: 会话 ID（必填）\n- limit: 返回消息数量上限（默认 50，最大 100）\n- before: 分页游标，指定上一页最后一条消息的 ID，用于加载更多历史消息\n\n返回：\n- 消息列表（倒序：最新消息在前）",
        "operationId": "get_chat_history_api_chat_history_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_MessageResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/sessions": {
      "get": {
        "tags": [
          "聊天会话管理"
        ],
        "summary": "List Sessions",
        "description": "列出当前用户的所有会话（按最后消息时间倒序）\n\n参数说明：\n- skip: 分页偏移量（默认 0）\n- limit: 每页数量（默认 20，最大 100）\n\n返回：\n- 会话列表（包含会话 ID、标题、状态、最后消息时间、消息数量）",
        "operationId": "list_sessions_api_sessions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "分页偏移量",
              "default": 0,
              "title": "Skip"
            },
            "description": "分页偏移量"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_SessionListItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}": {
      "get": {
        "tags": [
          "聊天会话管理"
        ],
        "summary": "Get Session",
        "description": "获取单个会话的详细信息\n\n参数说明：\n- session_id: 会话 ID\n\n返回：\n- 会话详细信息（包括标题、状态、开始时间、结束时间、最后消息时间）",
        "operationId": "get_session_api_sessions__session_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SessionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "聊天会话管理"
        ],
        "summary": "Update Session",
        "description": "更新会话信息（标题 / 状态 / 结束时间）\n\n参数说明：\n- session_id: 会话 ID\n- title: 新的会话标题（可选）\n- status: 新的会话状态（active/closed，可选）\n- ended_at: 会话结束时间（可选）\n\n使用场景：\n- 修改会话标题\n- 手动关闭会话\n- 标记会话已结束",
        "operationId": "update_session_api_sessions__session_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SessionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "聊天会话管理"
        ],
        "summary": "Delete Session",
        "description": "删除会话及其所有消息（软删除或硬删除）\n\n参数说明：\n- session_id: 会话 ID\n\n注意事项：\n- 删除会话会同时删除该会话下的所有消息\n- 请谨慎操作，删除后无法恢复\n- 只能删除属于自己的会话",
        "operationId": "delete_session_api_sessions__session_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponse_ChatResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[ChatResponse]"
      },
      "ApiResponse_List_MessageResponse__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[MessageResponse]]"
      },
...
      "ApiResponse_List_SessionListItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SessionListItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[SessionListItem]]"
      },
...
      "ApiResponse_SessionResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[SessionResponse]"
      },
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "ChatRequest": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "会话ID，为空则创建新会话"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "用户消息内容"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatRequest",
        "description": "发送聊天消息请求"
      },
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id",
            "description": "会话ID"
          },
          "message": {
            "$ref": "#/components/schemas/MessageResponse",
            "description": "助手的回复消息"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "ChatResponse",
        "description": "聊天响应（单条回复）"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryRequest": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id",
            "description": "会话ID"
          },
          "limit": {
            "type": "integer",
            "maximum": 100,
            "minimum": 1,
            "title": "Limit",
            "description": "返回消息数量上限",
            "default": 50
          },
          "before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before",
            "description": "分页游标：上一页最后一条消息的ID"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "HistoryRequest",
        "description": "获取会话历史请求"
      },
      "MessageResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id",
            "description": "所属会话ID"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role",
            "description": "消息角色"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "消息内容"
          },
          "meta_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Data",
            "description": "附加元数据"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "消息ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "消息发送时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "消息更新时间"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "role",
          "content",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "MessageResponse",
        "description": "消息信息响应"
      },
...
      "SessionListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "会话ID"
          },
          "session_uuid": {
            "type": "string",
            "title": "Session Uuid",
            "description": "会话UUID"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "会话标题"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "会话状态"
          },
          "last_message_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Message At",
            "description": "最后消息时间"
          },
          "message_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Count",
            "description": "会话消息总数"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_uuid",
          "status",
          "last_message_at"
        ],
        "title": "SessionListItem",
        "description": "会话列表项响应"
      },
      "SessionResponse": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "关联的用户ID，为空表示未登录会话"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "会话标题"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "closed"
            ],
            "title": "Status",
            "description": "会话状态",
            "default": "active"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "会话ID"
          },
          "session_uuid": {
            "type": "string",
            "title": "Session Uuid",
            "description": "会话UUID"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "会话开始时间"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At",
            "description": "会话结束时间"
          },
          "last_message_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Message At",
            "description": "最后消息时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_uuid",
          "started_at",
          "last_message_at"
        ],
        "title": "SessionResponse",
        "description": "会话信息响应"
      },
      "SessionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "新的会话标题"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "closed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "新的会话状态"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At",
            "description": "会话结束时间"
          }
        },
        "type": "object",
        "title": "SessionUpdate",
        "description": "更新会话请求"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/users/token"
          }
        }
      }
    }
  }
}